livesitter/backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
flask-cors==4.0.0